Review: Code restructure.
